refactor: Componentize submission actions button
